next-version/gateway/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.27.0

# RAG / Vector DB
chromadb>=0.4.18
openai>=1.3.0

# Text embeddings (local fallback)
sentence-transformers>=2.2.2

# Multimodal (CLIP) — optional, install for image embeddings
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
ftfy>=6.1.0
